refactor: move HTTPClientConfig and URL into util/config, fix import error in httputil
